requirements.txt
requests==2.17.3
tweepy==3.6.0
kafka-python==1.4.2
